Kernel: Fix race in waitid
This is similar to 29167a4e20c4b4e80f7147f425f71ee5d09fca1e
and 9aff77ca65368ed6402615295997056d83ce4689.

The crux is that wait verifies that the outvalue (siginfo* infop)
is writable *before* waiting, and writes to it *after* waiting.
In the meantime, a concurrent thread can make the output region
unwritable, e.g. by deallocating it.